@@ -6,46 +6,39 @@
   "dependencies": {
     "chartist": "0.9.8",
     "classnames": "2.2.5",
-    "es6-promise": "3.2.1",
-    "fetch-on-rest": "1.7.1",
-    "lodash": "4.13.1",
-    "react": "15.1.0",
-    "react-addons-test-utils": "15.1.0",
-    "react-bootstrap": "0.29.5",
-    "react-dom": "15.1.0",
-    "react-router": "2.5.1",
-    "urijs": "1.18.1",
+    "es6-promise": "3.3.1",
+    "fetch-on-rest": "2.0.0",
+    "lodash": "4.16.4",
+    "react": "15.3.2",
+    "react-addons-test-utils": "15.3.2",
+    "react-bootstrap": "0.30.5",
+    "react-dom": "15.3.2",
+    "react-router": "2.8.1",
+    "urijs": "1.18.2",
     "whatwg-fetch": "1.0.0"
   },
   "devDependencies": {
-    "babel-eslint": "^6.1.0",
-    "babel-jest": "^13.0.0",
-    "babel-loader": "6.2.4",
-    "babel-polyfill": "^6.9.1",
-    "babel-preset-es2015": "6.9.0",
-    "babel-preset-jest": "^13.0.0",
-    "babel-preset-react": "6.5.0",
-    "eslint": "^2.13.1",
-    "eslint-plugin-react": "5.2.2",
+    "babel-core": "^6.26.0",
+    "babel-eslint": "^7.2.3",
+    "babel-jest": "^16.0.0",
+    "babel-loader": "6.2.5",
+    "babel-polyfill": "^6.26.0",
+    "babel-preset-es2015": "6.16.0",
+    "babel-preset-jest": "^16.0.0",
+    "babel-preset-react": "6.16.0",
+    "eslint": "^3.19.0",
+    "eslint-plugin-react": "6.4.1",
     "exports-loader": "0.6.3",
-    "http-proxy": "1.14.0",
-    "jest-cli": "^13.0.0",
-    "jest-webpack-alias": "2.3.0",
-    "webpack": "1.13.1"
+    "http-proxy": "1.15.1",
+    "jest-cli": "^16.0.1",
+    "jest-webpack-alias": "3.3.0",
+    "webpack": "1.13.2"
   },
   "scripts": {
     "prestart": "webpack -d && (node server.js&)",
     "start": "webpack -d --watch",
     "build": "NODE_ENV=production webpack -p",
     "test": "jest --verbose"
-  },
-  "jest": {
-    "unmockedModulePathPatterns": [
-      "<rootDir>/node_modules/react/",
-      "<rootDir>/node_modules/react-dom/",
-      "<rootDir>/node_modules/react-addons-test-utils/",
-      "<rootDir>/node_modules/fbjs/"
-    ]
   },
   "repository": {
     "type": "git",
