@@ -1,16 +1,16 @@
 "use strict";
-var React = require('react');
-var isEqual = require('lodash/isEqual');
-var Utils = require('../components/utils.js');
-var Api = require('../api.js');
-var Misc = require('./misc.jsx');
-var Peers = require('./peers.jsx');
-var Results = require('./results.jsx');
-var ScrollBar = require('./scrollbar.jsx');
-var CompanyRatios = require('./ratios.jsx');
-var PriceChart = require('./pricechart.jsx');
-var QuickRatios = require('./quickratios.jsx');
-var AddCompare = require('./compare.jsx');
+import React from 'react'
+import isEqual from 'lodash/isEqual'
+import {setTitle} from '../components/utils.js'
+import Api from '../api.js'
+import {CompanyHeader, Ranges, Analysis, Announcements, AnnualReports, CompanyRatings} from './misc.jsx'
+import Peers from './peers.jsx'
+import Results from './results.jsx'
+import ScrollBar from './scrollbar.jsx'
+import CompanyRatios from './ratios.jsx'
+import PriceChart from './pricechart.jsx'
+import QuickRatios from './quickratios.jsx'
+import AddCompare from './compare.jsx'
 
 
 var Company = React.createClass({
@@ -37,7 +37,7 @@ var Company = React.createClass({
     var exc = params.exchange_code;
     var con = params.consolidated;
     Api.get(Api.company(exc, con)).then(function(response) {
-      Utils.setTitle(response.name);
+      setTitle(response.name);
       this.setState({
         company: response,
         comparisons: []
@@ -66,8 +66,9 @@ var Company = React.createClass({
 
   handleStartCompare: function() {
     var company = this.state.company
-    Api.get(Api.cid(company.id, 'comparison')).then(response => {
-      this.setState({comparisons: [response]});
+    var data = {force: company.warehouse_set.result_type}
+    Api.get(Api.cid(company.id, 'comparison'), data).then(resp => {
+      this.setState({comparisons: [resp]});
     });
   },
 
@@ -94,7 +95,7 @@ var Company = React.createClass({
     if (!company)
       return this.renderLoading()
     return <div>
-      <Misc.CompanyHeader
+      <CompanyHeader
         company={company}
         favorites={this.state.favorites}
         handleFavorite={this.handleFavorite}
@@ -112,7 +113,7 @@ var Company = React.createClass({
       </section>
 
       <section id="analysis">
-        <Misc.Analysis analysis={company.warehouse_set.analysis} />
+        <Analysis analysis={company.warehouse_set.analysis} />
       </section>
 
       <section id="peers">
@@ -130,7 +131,7 @@ var Company = React.createClass({
 
       <section id="annuals">
         <Results report="annual" company={company} comparisons={comparisons} />
-        <Misc.Ranges warehouse_set={company.warehouse_set} />
+        <Ranges warehouse_set={company.warehouse_set} />
       </section>
 
       <section id="balancesheet">
@@ -143,17 +144,17 @@ var Company = React.createClass({
 
       <section id="reports" className="hidden-print row">
         <div className="col-sm-12 col-lg-6">
-          <Misc.Announcements bse_code={company.bse_code} announcement_set={company.announcement_set} />
+          <Announcements bse_code={company.bse_code} announcement_set={company.announcement_set} />
         </div>
-        <div className="col-sm-4 col-lg-3" ng-if="vm.company.annualreport_set.length > 0">
-          <Misc.AnnualReports annualreport_set={company.annualreport_set} />
+        <div className="col-sm-4 col-lg-3">
+          <AnnualReports annualreport_set={company.annualreport_set} />
         </div>
-        <div className="col-sm-8 col-lg-3" ng-if="vm.company.companyrating_set.length > 0">
-          <Misc.CompanyRatings companyrating_set={company.companyrating_set} />
+        <div className="col-sm-8 col-lg-3">
+          <CompanyRatings companyrating_set={company.companyrating_set} />
         </div>
       </section>
     </div>
   }
 });
 
-module.exports = Company;
+export default Company
