@@ -1,6 +1,6 @@
 "use strict";
-var React = require('react');
-var Modal = require('./modal.jsx');
+import React from 'react'
+import Modal from './modal.jsx'
 
 
 var LoginModal = React.createClass({
@@ -25,4 +25,4 @@ var LoginModal = React.createClass({
 });
 
 
-module.exports = LoginModal;
+export default LoginModal
