@@ -1,12 +1,12 @@
 "use strict";
-var React = require('react');
-var ManageColumns = require('../modals/columns.modal.jsx');
-var Confirm = require('../components/confirm.jsx');
-var Notify = require('../components/notify.jsx');
-var api = require('../api.js');
-var Utils = require('../components/utils.js');
-var UserTable = require('./table.jsx');
-var ScreenBase = require('./base.jsx');
+import React from 'react'
+import ManageColumns from '../modals/columns.modal.jsx'
+import Confirm from '../components/confirm.jsx'
+import Notify from '../components/notify.jsx'
+import api from '../api.js'
+import {setTitle} from '../components/utils.js'
+import UserTable from './table.jsx'
+import ScreenBase from './base.jsx'
 
 
 class Screen extends React.Component {
@@ -48,7 +48,7 @@ class Screen extends React.Component {
   fetchResults(screenId, params) {
     var url = '/screens/' + screenId + '/';
     return api.get(url, params).then(resp => {
-      Utils.setTitle(resp.name);
+      setTitle(resp.name);
       this.setState({screen: resp, errors: false});
     }, resp => {
       this.setState({errors: resp, screen: false});
@@ -120,4 +120,4 @@ Screen.contextTypes = {
 }
 
 
-module.exports = Screen;
+export default Screen
