@@ -1,11 +1,9 @@
 "use strict";
-/* global require, window, jest */
+import getLastWord from '../cursor.js'
 
-jest.dontMock('../cursor.js');
 
 describe('test GetLastWord', function() {
   it('gets correct last words', function() {
-    var getLastWord = require('../cursor.js');
     var sentence;
     sentence = "Book value";
     expect(getLastWord(sentence)).toBe('Book value');
