@@ -1,14 +1,11 @@
 'use strict';
-/* global jest, require */
-jest.dontMock('../login.jsx');
-
+import React from 'react'
+import ReactDOM from 'react-dom'
+import TestUtils from 'react-addons-test-utils'
+import Login from '../login.jsx'
 
 describe('Login tests', function(){
   it('should render the login box', function() {
-    var React = require('react');
-    var ReactDOM = require('react-dom');
-    var TestUtils = require('react-addons-test-utils');
-    var Login = require('../login.jsx');
     var login = TestUtils.renderIntoDocument(
       <div><Login /></div>
     );
